review

+++ src/main/java/com/napier/sem/Capital.java
@@ -3,7 +3,7 @@ package com.napier.sem;
 public class Capital {
     private String name;
     private String country;
-    private int population;
+    private long population;
 
     public Capital(String name, String country, int population) {
         this.name = name;
@@ -19,7 +19,7 @@ public class Capital {
         return country;
     }
 
-    public int getPopulation() {
+    public long getPopulation() {
         return population;
     }
 }
